Add kickbar system; Switch sway to VictorMono
New system to force a repaint. Since the bar only repaints once per second, it was annoying to not have instant feedback on my volume changes and such. The new system runs `kickbar` each time it does something that could affect the bar, which traps USR1 in the status_command to repaint instantly.